fix:changed the position of view button
